Prevent accidental commiting of dev output in the future
When previewing the site locally (via "jekyll serve"), the domains
are changed away from that is in _config, and use localhost instead.

When building the site on its own (via "jekyll build") the
configured production domain is used.

Avoid mistakenly committing the output of the served preview by making
them go to different directories.

This uses the same approach as the performance/docroot.git repository.

Follows-up I91df03962c0 and Id7692ce6.

Change-Id: I20dd68a8e10a9a75d4a7d9ed8b06bd4fa41a1d3a